+++ Package.swift
@@ -8,7 +8,7 @@ let package = Package(
     name: "IBM Verify",
     defaultLocalization: "en",
     platforms: [
-        .iOS(.v15)
+        .iOS(.v17)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
@@ -27,9 +27,6 @@ let package = Package(
         .library(
             name: "MFA",
             targets: ["MFA"]),
-        .library(
-            name: "DC",
-            targets: ["DC"]),
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -108,31 +105,13 @@ let package = Package(
             dependencies: ["Core", "Authentication"],
             path: "Sources/mfa",
             resources: [
-                .copy("README.md")
+                .copy("README.md"),
+                .process("Resources")
             ]),
         .testTarget(
             name: "MFATests",
             dependencies: ["MFA", "Core", "Authentication"],
             path: "Tests/MFATests",
-            resources: [
-                .copy("Files")
-            ],
-            linkerSettings: [
-              .linkedFramework(
-                "XCTest",
-                .when(platforms: [.iOS])),
-            ]),
-        .target(
-            name: "DC",
-            dependencies: ["Core", "Authentication"],
-            path: "Sources/dc",
-            resources: [
-                .copy("README.md")
-            ]),
-        .testTarget(
-            name: "DCTests",
-            dependencies: ["DC", "Core", "Authentication"],
-            path: "Tests/DCTests",
             resources: [
                 .copy("Files")
             ],