+++ Sources/dc/WalletDescriptor.swift
@@ -1,9 +1,0 @@
-//
-// Copyright contributors to the IBM Verify Digital Credentials SDK for iOS project
-//
-
-import Foundation
-
-/// An interface that defines a wallet.
-public protocol WalletDescriptor: Codable {
-}